Add files via upload
Why: ran into the problem of not being able to stop it from regenerating the entire myLibrary array each time a new array object was added. The solution to this was to remove the forEach method and instead use a for of loop, with an if else statement to check if the object entered in the form was the last object in the array. 

Also included in this commit: 
  - Added the reset method to the form when the submit button was clicked.